Convert single file cartography.util to a directory
