+++ cartography/intel/aws/__init__.py
@@ -16,9 +16,9 @@ from . import redshift
 from . import resourcegroupstaggingapi
 from . import route53
 from . import s3
-from cartography.util import run_analysis_job
-from cartography.util import run_cleanup_job
-from cartography.util import timeit
+from cartography.util.jobs import run_analysis_job
+from cartography.util.jobs import run_cleanup_job
+from cartography.util.metrics import timeit
 
 logger = logging.getLogger(__name__)
