+++ cartography/intel/aws/ec2/__init__.py
@@ -9,7 +9,7 @@ from .security_groups import sync_ec2_security_groupinfo
 from .tgw import sync_transit_gateways
 from .vpc import sync_vpc
 from .vpc_peering import sync_vpc_peering
-from cartography.util import timeit
+from cartography.util.metrics import timeit
 
 logger = logging.getLogger(__name__)
