+++ cartography/intel/aws/ec2/instances.py
@@ -2,9 +2,9 @@ import logging
 import time
 
 from .util import get_botocore_config
-from cartography.util import aws_handle_regions
-from cartography.util import run_cleanup_job
-from cartography.util import timeit
+from cartography.intel.aws.util import aws_handle_regions
+from cartography.util.jobs import run_cleanup_job
+from cartography.util.metrics import timeit
 
 logger = logging.getLogger(__name__)
