+++ cartography/intel/aws/elasticsearch.py
@@ -4,10 +4,10 @@ import logging
 import botocore.config
 from policyuniverse.policy import Policy
 
+from cartography.intel.aws.util import aws_handle_regions
 from cartography.intel.dns import ingest_dns_record_by_fqdn
-from cartography.util import aws_handle_regions
-from cartography.util import run_cleanup_job
-from cartography.util import timeit
+from cartography.util.jobs import run_cleanup_job
+from cartography.util.metrics import timeit
 
 logger = logging.getLogger(__name__)
