+++ cartography/intel/aws/iam.py
@@ -4,8 +4,8 @@ import logging
 
 from cartography.intel.aws.permission_relationships import parse_statement_node
 from cartography.intel.aws.permission_relationships import principal_allowed_on_resource
-from cartography.util import run_cleanup_job
-from cartography.util import timeit
+from cartography.util.jobs import run_cleanup_job
+from cartography.util.metrics import timeit
 logger = logging.getLogger(__name__)
 
 # Overview of IAM in AWS
