+++ cartography/intel/aws/resourcegroupstaggingapi.py
@@ -1,9 +1,9 @@
 import logging
 from string import Template
 
-from cartography.util import aws_handle_regions
-from cartography.util import run_cleanup_job
-from cartography.util import timeit
+from cartography.intel.aws.util import aws_handle_regions
+from cartography.util.jobs import run_cleanup_job
+from cartography.util.metrics import timeit
 
 logger = logging.getLogger(__name__)
