+++ cartography/intel/aws/route53.py
@@ -1,7 +1,7 @@
 import logging
 
-from cartography.util import run_cleanup_job
-from cartography.util import timeit
+from cartography.util.jobs import run_cleanup_job
+from cartography.util.metrics import timeit
 
 logger = logging.getLogger(__name__)
