+++ cartography/intel/gcp/compute.py
@@ -7,8 +7,8 @@ from string import Template
 
 from googleapiclient.discovery import HttpError
 
-from cartography.util import run_cleanup_job
-from cartography.util import timeit
+from cartography.util.jobs import run_cleanup_job
+from cartography.util.metrics import timeit
 
 logger = logging.getLogger(__name__)
 InstanceUriPrefix = namedtuple('InstanceUriPrefix', 'zone_name project_id')
