+++ cartography/intel/gcp/crm.py
@@ -5,8 +5,8 @@ from string import Template
 
 from googleapiclient.discovery import HttpError
 
-from cartography.util import run_cleanup_job
-from cartography.util import timeit
+from cartography.util.jobs import run_cleanup_job
+from cartography.util.metrics import timeit
 
 logger = logging.getLogger(__name__)
