+++ cartography/intel/gcp/storage.py
@@ -3,8 +3,8 @@ import logging
 from googleapiclient.discovery import HttpError
 
 from cartography.intel.gcp import compute
-from cartography.util import run_cleanup_job
-from cartography.util import timeit
+from cartography.util.jobs import run_cleanup_job
+from cartography.util.metrics import timeit
 
 logger = logging.getLogger(__name__)
