+++ cartography/intel/github/repos.py
@@ -2,8 +2,8 @@ import logging
 from string import Template
 
 from cartography.intel.github.util import fetch_all
-from cartography.util import run_cleanup_job
-from cartography.util import timeit
+from cartography.util.jobs import run_cleanup_job
+from cartography.util.metrics import timeit
 
 
 logger = logging.getLogger(__name__)
