+++ cartography/intel/okta/groups.py
@@ -8,7 +8,7 @@ from okta.models.usergroup import UserGroup
 
 from cartography.intel.okta.utils import create_api_client
 from cartography.intel.okta.utils import is_last_page
-from cartography.util import timeit
+from cartography.util.metrics import timeit
 
 logger = logging.getLogger(__name__)
