+++ cartography/intel/okta/users.py
@@ -3,7 +3,7 @@ import logging
 
 from okta import UsersClient
 
-from cartography.util import timeit
+from cartography.util.metrics import timeit
 
 
 logger = logging.getLogger(__name__)
